[shuffle] create dev_api_client
Closes: #9846
