+++ shuffle/cli/src/lib.rs
@@ -3,6 +3,7 @@
 
 pub mod account;
 pub mod deploy;
+pub mod dev_api_client;
 pub mod new;
 pub mod shared;
 pub mod test;
